Add support for locale specific Selenium tests
Add static page locale information to FinancePages object and ensure clicking
the links in the page footer record this information when the locale is
changed. Add handlers for "Given page is in Welsh" and "Given page is in
English" cucumber directives to RequestForTransferSteps.

The RequestForTransferSteps validation_summary_should_be_displayed method now
checks for different text based on the currently active locale.

See PEN-139 (AC2)