Force view of initial page to be in English
Prevents occasional test failures if the previous test left the locale in
Welsh and the current test is expecting English. This happens because
English is the default locale and tests do not explicitly set the locale
unless it needs changing to Welsh.

Also remove references to @pen139 tag which caused the Cucumber
test runner to only run the contributionsAndPayment feature files.